Cambiando el tipo de dato de lo que tiene que ver con el precio de float a decimal

+++ Shared/Models/ItemCarrito.cs
@@ -20,7 +20,7 @@ namespace Shared.Models
 
         public int Cantidad { get; set; }
 
-        public float SubTotal {  get; set; }
+        public decimal SubTotal {  get; set; }
 
 
         public Carrito? Carrito { get; set; }
